requirements.txt
python-dotenv>=0.9.9
openai
streamlit
pydantic
